Developed login acceptance tests
